+++ gatsby-browser.js
@@ -2,3 +2,4 @@ import './src/styles/global.scss';
 import './src/styles/mobile.scss';
 import './src/styles/desktop.scss';
 import './src/styles/tablet.scss';
+import 'react-loading-skeleton/dist/skeleton.css';
