+++ src/pages/marketplace/index.tsx
@@ -1,20 +1,18 @@
 import React, { useState } from 'react';
 import { HeadFC } from 'gatsby';
-import { Tabs, Pagination } from 'antd';
+import { Pagination } from 'antd';
 import { IToken } from 'types/token';
-import useSWR from 'swr';
+import useSWR, { mutate } from 'swr';
 import { fetcher } from 'services/fetcher';
 import configs from 'config/config';
 import { toast } from 'react-toastify';
 import Layout from 'components/layout';
-import CardNFT from 'components/card-nft';
-import CardNFTSkeleton from 'components/card-nft-skeleton';
-import CollectionSkeleton from 'components/collection-skeleton';
+import CardToken from 'components/marketplace/card-token';
+import CardTokenSkeleton from 'components/marketplace/card-token-skeleton';
 
 const LIMIT = 12;
 
 function ListTokens() {
-    const [tab, setTab] = useState(1);
     const [page, setPage] = useState(1);
 
     const search = new URLSearchParams({
@@ -29,8 +27,45 @@ function ListTokens() {
         },
     });
 
-    const handleChangeTabKey = async (id: string) => {
-        setTab(+id);
+    const ListTokens = () => {
+        if (isLoading) {
+            return (
+                <div className="tabpane-content">
+                    {Array.from(Array(12).keys()).map((_, index) => (
+                        <CardTokenSkeleton key={index} />
+                    ))}
+                </div>
+            );
+        }
+        if (data) {
+            if (data.total == 0) {
+                return (
+                    <div className="text-center">
+                        <h4 className="text-white">No data</h4>
+                        <button
+                            className="btn btn-dark btn-small m-auto"
+                            onClick={() => mutate(endpoint)}
+                        >
+                            Reload
+                        </button>
+                    </div>
+                );
+            } else {
+                return (
+                    <div className="tabpane-content">
+                        {data.data.map((token: IToken) => {
+                            return <CardToken key={token.id} tokenInfo={token} />;
+                        })}
+                    </div>
+                );
+            }
+        }
+        return (
+            <div className="text-center">
+                <h4 className="text-white">Loading failed! Please try again</h4>
+                <button onClick={() => mutate(endpoint)}>Reload</button>
+            </div>
+        );
     };
 
     return (
@@ -58,77 +93,19 @@ function ListTokens() {
                     </div>
                 </div>
             </div>
-            <Tabs
-                defaultActiveKey="1"
-                activeKey={String(tab)}
-                centered={true}
-                className="marketplace-tabs"
-                onChange={handleChangeTabKey}
-            >
-                <Tabs.TabPane
-                    className="tabpane"
-                    tab={
-                        <>
-                            <span className="tabpane-title">
-                                NFTs <div className="tabpane-count">{data?.total ?? 0}</div>
-                            </span>
-                        </>
-                    }
-                    key="1"
-                >
-                    {tab === 1 && (
-                        <>
-                            <div className="tabpane-content">
-                                {!isLoading && data && data.data ? (
-                                    data.data.map((token: IToken) => {
-                                        return (
-                                            <CardNFT
-                                                key={token.id}
-                                                tokenInfo={token}
-                                                isLoading={isLoading}
-                                            />
-                                        );
-                                    })
-                                ) : (
-                                    <>
-                                        <CardNFTSkeleton />
-                                        <CardNFTSkeleton />
-                                        <CardNFTSkeleton />
-                                    </>
-                                )}
-                            </div>
-                            <Pagination
-                                current={page}
-                                pageSize={LIMIT}
-                                total={data?.total ?? 0}
-                                onChange={setPage}
-                            />
-                        </>
-                    )}
-                </Tabs.TabPane>
-                <Tabs.TabPane
-                    className="tabpane"
-                    tab={
-                        <span className="tabpane-title">
-                            Collections <div className="tabpane-count">67</div>
-                        </span>
-                    }
-                    key="2"
-                >
-                    {tab === 2 && (
-                        <>
-                            <div className="tabpane-content">
-                                <CollectionSkeleton />
-                            </div>
-                            <Pagination
-                                onChange={setPage}
-                                current={page}
-                                total={data?.total ?? 0}
-                            />
-                        </>
-                    )}
-                </Tabs.TabPane>
-            </Tabs>
+
+            <div className="min-h-screen">
+                <ListTokens />
+            </div>
+
+            {data && data.total > 0 && (
+                <Pagination
+                    current={page}
+                    pageSize={LIMIT}
+                    total={data?.total ?? 0}
+                    onChange={setPage}
+                />
+            )}
         </>
     );
 }
