+++ src/styles/desktop.scss
@@ -194,7 +194,7 @@
     }
 
     .browse-marketplace {
-        @apply px-[115px] py-[80px];
+        @apply px-[115px] py-[40px];
     }
 
     .browse-marketplace-title {
